Draft replacement v4 example for Annex 2 of INSPIRE mapping to GeoSciML cookbook
This is a GSML collection with member examples of nearly every type required by INSPIRE.
Current version does not validate because of problem with borehole schema (Borehole not in substitution group to be a member.)
Also the items in member elements have some extra containing elements because GSMLitem union class not turned into XSD properly by ShapeChange at the moment. Will be fixed but more a cosmetic change to serialisation.
Have included an entire MD_Metadata element as with the v3.2 cookbook example. The flexibility of gml:metaDataProperty means that potentially we could just have a subset of elements required to convey INSPIRE properties without the extra MD_Metadata required properties but this needs to be discussed.

git-svn-id: https://svnserv.csiro.au/svn/ext/GeoSciML@4787 910a3654-de2c-0410-bd95-d62e98ccebcd