Check-in: 3/07/2015 5:31:50 PM
Created leaf folders for GeologicEvent and GeologicStructure under GeoSciMLBasic. This will not affect the schemas that are generated for GeoSciMLBasic, but make it a lot easier to document and organise the UML (the GeologyBasic leaf was very large and it was proving very hard to create useful UML diagrams). Created new summary and context UML diagrams. (Still more to do.)
Replaced "CGI" prefixes on GSML data types with "GSML" prefix.  Reflects better on an OGC standard.
Changed several erroneous attribute links and imports from SWECommon v1 to v2 in various packages. 
Removed obsolete imports of ISO19115-2 in various packages (now using 19115-3). 
Regenerated package dependency diagrams.
Removed the GeologicDateEstimate class. Since the refactoring of Basic and Extended packages, this remnant v3.2 class was orphaned - it was not being used by any part of the v4 model, and was effectively covered by other elements in GeoSciMLBasic:GeologicalEvent and in the GeologicTime schema.

git-svn-id: https://svnserv.csiro.au/svn/ext/GeoSciML@4857 910a3654-de2c-0410-bd95-d62e98ccebcd